more DAS and ARR confusion
ughhhhh
Wiki measures DAS and ARR in milliseconds, I'm not sure how the games do it, but given the customization I'm assuming it's millisecond based in jstris.
The entire task system I set up to handle asynchronous millisecond-based events for DAS is there and working, but my difficulties with the system aren't over yet.